New Tau Table version 2, and support code
This commit includes the version 2 pexit and cdf tables, which are a
concatenation along the beta_rad axis of the below-1-degree tables from
v1 and the entirety of the v0 tables. Additional code was added,
including the new `grid_concatenate` function in utils.grid and the
creation and test scripts in the data directory. The default tau table
file version has been upgraded to 2 in the configuration code.